1.37.24 (5.3.6f1)
Build-Id: 2b381d9e-f208-4758-b071-a963f16905ca
Build-Sha256: 92bd19c7f1f134817bbb1c5bd0b1fc75344dbfea2cf24db4c0148ede63eea6da

+++ Assembly-CSharp/AsyncWWWManager.cs
@@ -21,7 +21,7 @@ public static class AsyncWWWManager
 
 	private static int maxRequests = 4;
 
-	public static readonly float[] RetryTimeouts = new float[3] { 30f, 20f, 10f };
+	public static readonly int[] RetryTimeouts = new int[3] { 30, 20, 10 };
 
 	private static Dictionary<WWWRequestPriority, Queue<AsyncWebRequest>> requests = new Dictionary<WWWRequestPriority, Queue<AsyncWebRequest>>
 	{
